Implement Signin and Signup activity
