+++ app/build.gradle
@@ -34,4 +34,5 @@ dependencies {
     testCompile 'junit:junit:4.12'
 }
 
+
 apply plugin: 'com.google.gms.google-services'
