+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="edu.brandeis.cs.bummer">
+
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
@@ -13,10 +14,12 @@
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".EmailPasswordActivity"></activity>
+        <activity android:name=".SigninActivity" />
+        <activity android:name=".SignupActivity"></activity>
     </application>
 
 </manifest>
