+++ app/src/main/java/edu/brandeis/cs/bummer/MainActivity.java
@@ -1,16 +1,18 @@
 package edu.brandeis.cs.bummer;
 
+import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
+import android.widget.Toast;
 
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
 public class MainActivity extends BaseActivity implements View.OnClickListener {
     private static final String TAG = "MainActivity";
+    private static final int SIGN_IN = 7;
     private FirebaseUser currentUser;
 
     // [START declare_auth]
@@ -34,13 +36,17 @@ public class MainActivity extends BaseActivity implements View.OnClickListener {
         if (currentUser != null) {
             Log.d(TAG, "current user = " + currentUser.getUid());
         } else {
-
+            signIn();
         }
+    }
+
+    private void signIn() {
+        startActivityForResult(new Intent(this, SigninActivity.class), SIGN_IN);
     }
 
     private void signOut() {
         mAuth.signOut();
-        startActivity(new Intent(this, EmailPasswordActivity.class));
+        signIn();
     }
 
     @Override
@@ -48,6 +54,21 @@ public class MainActivity extends BaseActivity implements View.OnClickListener {
         int i = v.getId();
         if (i == R.id.sign_out_button) {
             signOut();
+        }
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if(requestCode == SIGN_IN) {
+            if (resultCode == Activity.RESULT_OK) {
+                Log.d(TAG, "in ok");
+                Toast.makeText(this, "ok", Toast.LENGTH_SHORT).show();
+            }
+            if (resultCode == Activity.RESULT_CANCELED){
+                Toast.makeText(this, "canceled", Toast.LENGTH_SHORT).show();
+            }
         }
     }
 }
